requirements.txt
glcontext==3.0.0
moderngl==5.12.0
numpy==2.0.2
pygame==2.6.1
PyYAML==6.0.1
